+++ modules/site/mod_web357_random_recipe/services/provider.php
@@ -1,0 +1,25 @@
+<?php
+
+/**
+ * @version    CVS: 1.0.0
+ * @package    Mod_Web357_Random_Recipe
+ * @author     Web357 Dev <careers@web357.com>
+ * @copyright  2025 Web357.com
+ * @license    GNU General Public License version 2 or later; see LICENSE.txt
+ */
+\defined('_JEXEC') or die;
+
+use Joomla\CMS\Extension\Service\Provider\HelperFactory as HelperFactoryServiceProvider;
+use Joomla\CMS\Extension\Service\Provider\Module as ModuleServiceProvider;
+use Joomla\CMS\Extension\Service\Provider\ModuleDispatcherFactory as ModuleDispatcherFactoryServiceProvider;
+use Joomla\DI\Container;
+use Joomla\DI\ServiceProviderInterface;
+
+return new class () implements ServiceProviderInterface {
+
+    public function register(Container $container): void
+    {
+        $container->registerServiceProvider(new ModuleDispatcherFactoryServiceProvider('\\Webtest\\Module\\Web357RandomRecipe'));
+        $container->registerServiceProvider(new ModuleServiceProvider());
+    }
+};